Support declarative datasets in datacompiler
- Use the dataset descriptions to define compiler handlers
- Reduce common.py to a Compiler with datasets_description set to
  source/datasets/idkbse.ttl
