+++ lxltools/datacompiler.py
@@ -24,8 +24,9 @@ CACHE_SPAQRL_BASE = 'urn:x-cache:sparql:'
 
 class Compiler:
 
-    def __init__(self,
+    def __init__(self, *,
                  base_dir=None,
+                 datasets_description=None,
                  dataset_id=None,
                  tool_id=None,
                  created=None,
@@ -33,6 +34,7 @@ class Compiler:
                  record_thing_link='mainEntity',
                  system_base_iri=None,
                  union='all.jsonld.lines'):
+        self.datasets_description = datasets_description
         self.datasets = {}
         self.current_ds_resources = set()
         self.base_dir = tracked_path_type(self.current_ds_resources)(base_dir)
@@ -46,6 +48,9 @@ class Compiler:
         self.union = union
         self.current_ds_file = None
         self.no_records = False
+
+        if datasets_description:
+            self._handlers_from_datasets_description(datasets_description)
 
     def main(self):
         argp = argparse.ArgumentParser(
@@ -372,6 +377,65 @@ class Compiler:
     def construct(self, sources, query=None):
         return _construct(self, sources, query)
 
+    def _handlers_from_datasets_description(self, description_path):
+        g = Graph().parse(self.path(description_path), format='turtle')
+        data = self.to_jsonld(g)
+        for ds in data['@graph']:
+            if ds.get('@type') != 'Dataset' or 'uriSpace' not in ds:
+                continue
+            self.dataset(_make_handler(self, ds))
+
+
+def _make_handler(compiler, ds):
+    def dataset_handler():
+        source = _digest_source_data(ds['sourceData'])
+        if len(src := source['source']) == 1 and 'query' in src[0]:
+            source = src[0]
+
+        graph = compiler.construct(
+            sources=source.get('source', []),
+            query=source.get('query')
+        )
+
+        ztime = ds['created']['@value'].replace('+00:00', 'Z')
+        return ds.get('uriSpace'), ztime, graph
+
+    dataset_handler.__name__ = ds['@id'].rsplit('/', 1)[-1]
+
+    return dataset_handler
+
+
+def _digest_source_data(src):
+    source = {}
+    unhandled = False
+
+    if 'dataQuery' in src:
+        assert 'query' not in source
+        source['query'] = src['dataQuery']['uri']
+    elif '@id' in src:
+        assert 'source' not in source
+        source['source'] = str(src['@id']) # TODO: bug in rdflib; URIRef in the json-ld
+    elif 'uri' in src:
+        instruct = 'result' if 'sourceData' in src else 'source'
+        assert instruct not in source
+        source[instruct] = src['uri']
+    else:
+        unhandled = True
+
+    if 'representationOf' in src:
+        instruct = 'dataset'
+        assert instruct not in source
+        source[instruct] = src['representationOf']['@id']
+        unhandled = False
+
+    for part in _aslist(src.get('sourceData')):
+        source.setdefault('source', []).append(_digest_source_data(part))
+        unhandled = False
+
+    assert not unhandled
+
+    return source
+
 
 def last_modified_ms(fpaths):
     time_stamps = [f.stat().st_mtime for f in fpaths]
@@ -469,3 +533,7 @@ def _construct(compiler, sources, query=None):
         for spo in result:
             g.add(spo)
         return g
+
+
+def _aslist(o):
+    return o if isinstance(o, list) else [] if o is None else [o]
